requirements.txt
pypdf2>=3.0.0
pymupdf>=1.23.0
openai>=1.0.0
anthropic>=0.7.0
markdown>=3.4.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pillow>=10.0.0
numpy>=1.24.0
